Fix: selection actions not working (#332)
