+++ packages/ui/src/features/store/objects/selection/ObjectsActionSpec.ts
@@ -11,7 +11,6 @@ export interface ObjectsActionSpec {
     component: ActionComponentType;
     hideInList?: boolean;
     destructive?: boolean;
-    isWorkflow?: boolean;
 }
 
 export interface ObjectsActionParams {
