+++ packages/ui/src/features/store/objects/selection/index.ts
@@ -1,5 +1,4 @@
 export * from "./actions";
-export * from "./ActionsRenderer";
 export * from "./ObjectsActionContext";
 export * from "./ObjectsActionSpec";
 export * from "./SelectionActions";
